Remove shortcode performance plugin test class
